[refactor] separate api calls
